Revert "proyecto de validacion de formularios en thymeleaf"
This reverts commit b184a2a12e3ed701d8de3c0f2377c119aeeb408e.

+++ com.jaimebackend.primeraweb/src/main/java/com/jaimebackend/primeraweb/service/PersonaService.java
@@ -16,5 +16,5 @@ public interface PersonaService {
 
     void eliminarPersona(Long id);
 
-    int contarPersonas();
+    long contarPersonas();
 }
